+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Ensure Afterpay sends a non-empty Stripe billing and shipping name when one is available.
+
 ## [2.0.1] - 2026-04-24
 
 ### Fixed